Split parametric_ops.py into 3 files
