+++ pennylane/ops/qubit/special_unitary.py
@@ -22,7 +22,7 @@ from itertools import product
 import numpy as np
 import pennylane as qml
 from pennylane.operation import AnyWires, Operation
-from pennylane.ops.qubit.parametric_ops import PauliRot
+from pennylane.ops.qubit.parametric_ops_multi_qubit import PauliRot
 
 _pauli_matrices = np.array(
     [[[1, 0], [0, 1]], [[0, 1], [1, 0]], [[0, -1j], [1j, 0]], [[1, 0], [0, -1]]]
